+++ src/app/(main)/chat/(workspace)/_layout/Desktop/ChatHeader/Main.tsx
@@ -3,11 +3,11 @@
 import { ActionIcon, Avatar, ChatHeaderTitle } from '@lobehub/ui';
 import { Skeleton } from 'antd';
 import { PanelLeftClose, PanelLeftOpen } from 'lucide-react';
-import Link from 'next/link';
-import { memo } from 'react';
+import { Suspense, memo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
+import { useInitAgentConfig } from '@/app/(main)/chat/(workspace)/_layout/useInitAgentConfig';
 import { useOpenChatSettings } from '@/hooks/useInterceptingRoutes';
 import { useGlobalStore } from '@/store/global';
 import { useSessionStore } from '@/store/session';
@@ -17,6 +17,8 @@ import Tags from './Tags';
 
 const Main = memo(() => {
   const { t } = useTranslation('chat');
+
+  useInitAgentConfig();
 
   const [init, isInbox, title, description, avatar, backgroundColor] = useSessionStore((s) => [
     sessionSelectors.isSomeSessionActive(s),
@@ -45,20 +47,19 @@ const Main = memo(() => {
   ) : (
     <Flexbox align={'flex-start'} gap={12} horizontal>
       {
-        <Link aria-label={t('agentsAndConversations')} href={'/chat'}>
-          <ActionIcon
-            icon={showSessionPanel ? PanelLeftClose : PanelLeftOpen}
-            onClick={() => {
-              const currentShowSessionPanel = useGlobalStore.getState().preference.showSessionPanel;
-              useGlobalStore.getState().updatePreference({
-                sessionsWidth: currentShowSessionPanel ? 0 : 320,
-                showSessionPanel: !currentShowSessionPanel,
-              });
-            }}
-            size="large"
-            title={t('agentsAndConversations')}
-          />
-        </Link>
+        <ActionIcon
+          aria-label={t('agentsAndConversations')}
+          icon={showSessionPanel ? PanelLeftClose : PanelLeftOpen}
+          onClick={() => {
+            const currentShowSessionPanel = useGlobalStore.getState().preference.showSessionPanel;
+            useGlobalStore.getState().updatePreference({
+              sessionsWidth: currentShowSessionPanel ? 0 : 320,
+              showSessionPanel: !currentShowSessionPanel,
+            });
+          }}
+          size="large"
+          title={t('agentsAndConversations')}
+        />
       }
       <Avatar
         avatar={avatar}
@@ -72,4 +73,17 @@ const Main = memo(() => {
   );
 });
 
-export default Main;
+export default () => (
+  <Suspense
+    fallback={
+      <Skeleton
+        active
+        avatar={{ shape: 'circle', size: 'default' }}
+        paragraph={false}
+        title={{ style: { margin: 0, marginTop: 8 }, width: 200 }}
+      />
+    }
+  >
+    <Main />
+  </Suspense>
+);