+++ src/libs/agent-runtime/deepseek/index.ts
@@ -1,0 +1,15 @@
+import { AgentRuntimeErrorType } from '../error';
+import { ModelProvider } from '../types';
+import { LobeOpenAICompatibleFactory } from '../utils/openaiCompatibleFactory';
+
+export const LobeDeepSeekAI = LobeOpenAICompatibleFactory({
+  baseURL: 'https://api.deepseek.com/v1',
+  debug: {
+    chatCompletion: () => process.env.DEBUG_DEEPSEEK_CHAT_COMPLETION === '1',
+  },
+  errorType: {
+    bizError: AgentRuntimeErrorType.DeepSeekBizError,
+    invalidAPIKey: AgentRuntimeErrorType.InvalidDeepSeekAPIKey,
+  },
+  provider: ModelProvider.DeepSeek,
+});